+++ packages/eas-cli/src/metadata/__tests__/fixtures/invalid.async.js
@@ -1,0 +1,3 @@
+module.exports = async function () {
+  return await Promise.resolve({ configVersion: -1 });
+};