+++ packages/eas-cli/src/metadata/__tests__/fixtures/invalid.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  configVersion: -1,
+};