+++ packages/eas-cli/src/metadata/__tests__/fixtures/store.async.js
@@ -1,0 +1,5 @@
+module.exports = async function () {
+  const config = require('./store.config.json');
+  config.apple.copyright = `${new Date().getFullYear()} ACME`;
+  return await Promise.resolve(config);
+};